refactor(TestSetterAndGetterTrait): Do not automatically set 'assert' to 'same' if value is an object
It was impossible to check a object values for equality without specifying the 'assert' method.
This could have been unintuitive in certain test situations - while the opposite case ( check value objects to be
the exact same) had not to be made explicit.

Now checking if two value objects are the same has to be made explicit by specifying the assert method.